Cancel Project Run if WorkingDirectory doesn't exist and report
Previously if a user had set their run configuration to use the "run in directory" with a path that did not exist, MonoDevelop would hang up upon attempting "Start Debugging" or "Start Without Debugging". It would not cancel or alert the user of why this was occurring which could be quite confusing to figure out.

Instead, during a call to OnExecuteCommand, the working directory is checked to see if it exists. If it does, it continues normally. If not it cancels the execution and reports a warning explaining what the problem is, indicating to the user where to go to fix the issue.