[Roslyn] Rework persistent storage location service  (#7057)
* [Workspace] Dispose of the base workspace at the end, otherwise it clears all the Solution data

* [Roslyn] Rework persistent storage location service

Brings it in line with VSWorkspace implementation.

We only support one main solution for persistence, the others
are ignored. So a secondary solution in the workspace will not use persistence.

Also, we do not handle the case where we do:
* Open Solution A
* Open Solution B
* Close Solution A

Instead of using persistence for B, we do not. We will use it for the next solution
that is opened.

Fixes VSTS #786109 - SQLitePersistentStorage not disposed on solution close and other issues

* [CodeFix] Simplify some task run, queue just one task, not two