Defensive programming in Dock Container
MERP / Dr Watson show this as one of the most frequent crashes in VSMac

```
    Fault:       Crash
    thread 0:
        IsManaged:   False
        NativeFrames:0
        Managed Frames:
            MonoDevelop.Ide.dll!DockContainer.UpdatePlaceholder+20
            MonoDevelop.Ide.dll!DockFrame.UpdatePlaceholder+0
            MonoDevelop.Ide.dll!DockItemTitleTab.OnMotionNotifyEvent+d0
            40f46e55-eb67-454a-b904-dc9e5644a3a1!6003c23+16 (UNKNOWN MVID)
```

which points to `placeholderWindow.AllowDocking = allowDocking;` as the line
that is failing.

It isn't obvious at all how placeholderWindow could ever be null at this point.
(I don't see access to the placeholderWindow field from threads other than the
UI), so I have just made the method more robust.