Fix VSTS 788692 - crash in PositionChanged_ITextCaret
Watson reports indicate a hard crash with this callstack:

           MonoDevelop.SourceEditor.dll!CaretImpl.PositionChanged_ITextCaret+ca
           MonoDevelop.SourceEditor.dll!CaretImpl.OnPositionChanged+1d

We're not sure of the repro or how it can happen but nothing critical is happening in this codepath (just updating the MultiSelectionBroker).

This has been a common source of trouble in the past, and we think it's better to have a slight risk of MultiSelectionBroker de-sync than a hard crash of the IDE with data loss etc.

Worst case that can happen is that IntelliSense will show in the wrong place or some other minor editor glitches. But we shouldn't crash here.