Packaging: Ensure IPackageInstallerService.PackageSources is thread-safe
For the new editor, we are importing Roslyn's suggestions code for the
LightBulb feature.

Unfortunately, this code currently expects to run on the main thread,
but then uses `Task.Run(...).Wait()` when fetching code fixes and
refactorings. Therefore, code fix providers can't call back into the
main thread, or else there will be deadlock.

The "Add Import" code fix provider calls
`IPackageInstallerService.PackageSources`. MonoDevelop's implementation
was blocking on the main thread. Remove this, add comments to prevent
problems in the future, and use an immutable list in
`PackageManagementWorkspace` to allow for thread-safe list lookups.