DelegateCompletionProvider: Do not invoke into main thread with new editor
This prevents a hang that could occur when invoking the quick fix
command. One of the quick fix sources apparently calls into completion,
but as we know from experience, Roslyn's suggestions code is very greedy
with the main thread, so `await Runtime.RunInMainThread` is no longer
okay in intellisense providers (with a few exceptions).

Old editor still gets old behavior to avoid regressing it.