Implemented new document/view architecture
Introduced the new DocumentController and related classes, which replaces
ViewContent and BaseView content. A document controller is a class that
implements the logic for loading, displaying and interacting with the contents
of a document. The FileDocumentController class is a controller base class
specialized in showing the content of files.

DocumentController supports extensions using the DocumentControllerClass.
Extensions can hook into controller operations and provide custom UI.

This commit also introduces the DocumentView class and derived classes.
A DocumentView is a component of the shell that can show the content provided
by a document controller. The DocumentViewContent subclass can show a Control,
and the DocumentViewContainer can show a set of views using tabs or splits.

Unlike the old ViewContent class, which created a Control, the
DocumentController class creates a DocumentView.

SdiWorkspaceWindow has changed to support the new DocumentView model.
However, all the interaction between DocumentView and the shell implementation
is now done through interfaces so that porting to a native toolkit will be easier in
the future, and also makes the code easier to unit test.

There are several GtkShellDocumentView* classes that implement the
functionality of DocumentView and related classes.

Document handling logic has been moved from Workbench to a new DocumentManager
class.

The Document class is not a subclass of DocumentContext anymore. It has
a DocumentContext property instead. All logic in charge of parsing documents
has been moved to the new RoslynDocumentExtension class, a document
controller extension that is created only when the document is parseable.