Implemented DocumentModel and derived classes
A model is the data being shown in a view. DocumentModel is the
base class for models that can be edited in a view.
FileModel is a base class for file-based models. TextFileModel is
the base class for models stored in text files. TextBufferFileModel
is a model that stores data in a TextBuffer.
DocumentModelRegistry is a service that can be used to share
and get shared models.