Treat ctrl+click like right-click in mouse processors
I don't believe this is the right way to fix this bug. If ctrl+click
should always be interpreted as a right-click, then implementers of
`ICocoaMouseProcessorEvents` should do these checks before deciding
which click event to raise.

Additionally, it seems like it would be preferable to let the framework
handle this by setting the NSMenu on the view or implementing
`menuWillOpen:`, etc.

Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/795494