MDMenuItem: avoid deprecated NSMenuItem.SetTitleWithMnemonic
Mnemonics have been deprecated as a concept since macOS 10.6. The API
has supported them but they stopped rendering and functioning then.

SetTitleWithMnemonic was formally deprecated in 10.13, even though
all the other APIs were deprecated in 10.6 and turned into noop.

Because the & character was used to indicate a mnemonic in macOS, these
characters would not actually render in the title of the menu item
when using SetTitleWithMnemonic.

Instead, just set the Title property, and discard the CommandInfo
mnemonic character entirely.