[Core] Handle missing language on saving shared assets project
If a shared assets project was missing the import for its projItems
then this could result in a null reference exception when the project
was saved.

 <Import Project="Shared.projitems" Label="Shared" />

This was because the language is not set if the above import is
missing.

Fixes VSTS #802484 Unable to save Xamarin.Android-Tests.sln after adding
an existing shared project