[Core] Handle a shared assets project type guid used for C# project
If the solution file is edited by hand and a shared assets project
entry was converted to a C# project file but the shared assets
project type guid is kept then the project would be loaded as a
shared assets project. To avoid this a check is made to see if the
project file extension is correct on reading the project file. If
not then an exception is thrown indicating there is a problem with
the project type guid. This avoids the project being treated as a
shared assets project incorrectly.