[Core.Tests] Add shared assets project tests
Added a test to ensure a null reference exception does not occur
on saving a shared assets project if the language is not defined.
Added a test to check an exception is thrown on loading a solution
that is using a shared assets project type guid for a C# project.