VSTS 801818: DefaultSourceEditorOptions: return false for WrapLines
The old editor apparently still attempts to do something when WrapLines
returned true. When we introduced word wrapping in the new editor, the
property changed from a boolean to a flags enum. For compatibility we
plumbed the old WrapLines property through to the new WordWrapStyle
property. This was a bad assumption and introduced VSTS #801818.

Apparently the old editor may have at some point tried to implement word
wrapping and has been respecting the preference all along but it has
been dormant for 6 years since it was disabled in the preferences UI.

The deprecated WrapLines property will again always return false, fully
disconnecting it from the preferences UI.

https://devdiv.visualstudio.com/DevDiv/_workitems/edit/801818