Implement MiscellaneousFilesWorkspace.
This workspace tracks open documents and in case an open document is not claimed by any other workspace, the misc workspace picks it up to provide basic Roslyn services such as classification.

The misc workspace uses the WorkspaceRegistration per open TextBuffer/SourceTextContainer to track when a document is added to or removed from any workspace.