CocoaTextViewContent: Implement GrabFocus
Put focus logic in `EmbeddedNSViewControl` (necessitated modifying
`Control` to allow overriding focus API). Realistically, focus is driven
from the view content and not the control, but this seemed safer and
cleaner.

`ViewContent.GrabFocus` is called when tabs are changed, etc. This
implementation focuses the GTK+ host widget, then attempts to focus the
text view. Because the GTK+ host widget removes our `NSView` from its
superview on unrealized/unmapped/etc, it may not have an `NSWindow` at
this time. In this case, rely on the new `HostControlMovedToWindow`
event.

Ideally, we would focus the `CocoaEditorGridView` (aka the native host),
but for some reason this is not working. Instead, we focus the text view
itself. This can be wrong, for example if the find view is meant to be
focused. We can work on this more in the future.

From this point, the proper way to ensure focus returns to us from other
IDE components is to call `GrabFocus` on the view content. This matches
behavior seen in the old editor.

Fixes: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/802462