[TextEditor] Support project reloads with new editor
With a C# file open in the new editor in another tab, and a .csproj
was edited and saved, a dialog would appear asking to reload the
changes:

The project 'Test' has been modified by an external application. Do
you want to reload it?

Some files may need to be reloaded, and editing status for those
files (such as the undo queue) will be lost.

This dialog was displayed because the new editor's ViewContent
did not indicate it supported project reloading. The IDE's
RootWorkspace checks the Document's ProjectReloadCapability to
see if all open documents can be reloaded automatically when the
project is reloaded. If any ViewContent indicates it does not have
full project reload support then a dialog is displayed. The old
editor was unaffected.

Fixes VSTS #803378 - editing csproj prompts to reload changes