[TextEditor] Fix null ref on reloading project with new editor open
On reloading the Document tries to initialize the old text editor's
extension chain causing a null reference exception.

System.NullReferenceException: Object reference not set to an instance of an object
  at MonoDevelop.Ide.Gui.Document.InitializeExtensionChain () [0x00001] in main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.Gui/Document.cs:726
  at MonoDevelop.Ide.Gui.Document.SetProject (MonoDevelop.Projects.Project project) [0x000bf] in main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.Gui/Document.cs:831
  at MonoDevelop.Ide.RootWorkspace.ReattachDocumentProjects (System.Collections.Generic.IEnumerable`1[T] closedDocs) [0x00054] in main/src/core/MonoDevelop.Ide/MonoDevelop.Ide/RootWorkspace.cs:553

Avoid this by checking if the Editor is null and only running the code
that is relevant for the new editor.