[Ide] Fix project not attached back to view on project reload
Causing a project or solution to be reloaded in the IDE, such as when
these files are changed externally, would not attach the reloaded
project back to the Document and its associated ViewContent. On
reloading the RootWorkspace finds the documents for the project
being reloaded. It then calls Document.SetProject passing a null.
Then after the reload the RootWorkspace then calls SetProject again
with a project it finds for the document if the document has no
Project associated. The problem was that the Document.SetProject
when passed null was not changing the project for the view content.
So the reloaded project was not being attached back to the document.

Fixing the above problem then caused a problem for the editor where
it would create an ad hoc project, since its project had been set to
null, and then the reloaded project would not be attached back to
the document. To avoid this a check is made to ensure the parser
thread, which creates an ad hoc document, is not run if the document
was not using an ad hoc project before when SetProject is called with
a null project.

Fixes VSTS #807754 - Reloading a project does not update the project
for the Document