[Ide] Fix progress dialog in front of confirmation dialog
Adding an existing folder, that does not exist inside the project and
has more than 10 files, was opening a progress monitor dialog in front
of the confirmation dialog asking how to copy the files into the
project. A change to the MessageDialogProgressMonitor affected its
parent causing it to appear above the confirmation dialog. To fix this
the MessageDialogProgressMonitor now has a constructor that takes the
parent window.

Fixes VSTS #807518 - New ASP.NET Core Empty > .NET Core 2.x project,
add Existing Folder "bootstrap-4.3.1-dist", "Adding files..." dialog
in the front.