Fixes VSTS Bug 806781: Beachball on solution search when it has large JavaScript files
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/806781

This does not really fix the problem - it shadows it a bit. The line
is 1.8MB long - highlighting that with the current text mate engine is
problematic.
However even sublime text refuses to try to highlight such long lines.
But it shows a pango rendering error as well which can't be fixed.

But it no longer breaks the VS4MAC editor in a dead-lock like
situation.
However may be a good test for the new editor.