[Ide] Fix navigation history skipping files
After selecting Go to definition you could not navigate back to the
original file since the navigation point was removed from the history.

This reverts part of 0544da29231d240a8cc15477ebfa64bda1b9e3e2 which
had an extra check to see if the current navigation point is a match
for the new navigation point.

The old editor behaviour now works as before. The new editor has a
problem where it adds a navigation point for the first line in a file
when it is opened but navigating back now works.

Fixes VSTS #820638 - Navigation history removes files from history