[Ide] Close open documents if parent directory is removed externally
Deleting a directory outside the IDE would not close documents that
were open in the IDE. Deleting individual files outside the IDE
that corresponded to open documents were closed correctly. The problem
was that the file service events for a directory being deleted do
not indicate a directory was removed so the existing logic was being
bypassed. To handle this, when a deletion event is processed, if no
exact filename match is found, then a check is made to see if the
deleted item matches a parent directory for any open view. If there
is a match then the document is closed.

Fixes VSTS #823237 Document not closed when parent directory removed