[Ide] Fix UI unresponsive after deleting multiple projects
Selecting two projects in the Solution pad, right clicking and
selecting Delete would result in the IDE being unresponsive, menus
were disabled and you could not quit. The problem was that after
the first project is deleted a confirmation dialog is displayed
to confirm the second project should be deleted. This confirmation
dialog was displayed before the progress monitor's message dialog
had been closed so it used this dialog as its parent. This resulted
in the confirmation dialog appearing and then disappearing. The IDE
was then waiting for the dialog to be closed but it is not available
anywhere. To prevent this the confirmation dialog has its parent
window set to be the main IDE window.

Fixes VSTS #827646 - Visual Studio hangs when deleting multiple
projects together.