[Ide] Use Gtk.TreePath to in-/validate the current status popup
To verify the target node of the currently shown status
tooltip we cache the Gtk.TreeIter and compare its path
to the current one. Since we actually already know the
Gtk.TreePath, there is no need to convert iters, we
can just compare the node path and avoid the loop
into the potentially slow store.GetPath().

Fixes VSTS #827706