[Mac] Set SharedApplication.HelpMenu properly
If SharedApplication.HelpMenu is not set, Cocoa tries to find
the best matching menu to attach the help search. This usually
fails for localized environments and sometimes even for english.

https://developer.apple.com/documentation/appkit/nsapplication/1428644-helpmenu:
"When the value is nil, AppKit installs Spotlight menu items on the
menu of its choosing. To suppress Spotlight help items altogether,
specify a menu that does not appear on the menu bar."

Fixes VSTS #832346