[Ide] Reload  most recently used projects first in type system
If the project info cache is available then reloading the project
info from the project system will be done with the recently used
projects being reloaded first. The recently used projects is based
on the documents that were open previously when the solution was
closed. This information is stored in the user prefs and is also
used to re-open these files on loading the solution so this is being
re-used here.