[Ide] Remove in-memory project info from cache when no longer needed
The project info cache stores information to disk when its Update
method is called. Once the initial information loaded from the cache
has been used the in-memory cache can be cleared for that project.
An alternative might be to keep the in-memory cache and then only
write out the project info cache to a file when the solution is
closed in the IDE.