[Ide] Load project info cache on background thread.
The project info cache was being loaded in the cache class
constructor when the associated workspace was created. This is done
on the UI thread when the solution is first loaded. To avoid this
the loading has been moved out of the constructor into a separate
Load method which is called by the type system before it attempts
to use the cache information.