Fix document controller initialization issues
OnLoad is now ensured to be called before a document is shown.
Fixed an issue that happened when a controller provides a view
that has been created by another controller. In that case the
controllers are chained, so that the original controller can
still get the events from the view.