[Ide] Add tests for project info cache
Tests:

Cache is created on loading the solution into the type system.
Cache is updated when the project is modified.
Cache is updated when a file is added externally to an SDK style
project.
Cache is used on initial load before MSBuild is used.
Cache is updated if it is out of date compared with the MSBuild
information.