[Core] Improve performance when looking for matching globs for new file
Building an ASP.NET Core project, that restores missing NPM modules
during the build, was causing the IDE to use a lot of CPU and the build
would seem to never end. The IDE was spending its time checking to
see if the created file should be automatically added to the project
and for an NPM restore this can involve thousands of files. The problem
is that the file system watcher in Mono, which is now used, fires file
creation events on thread pool threads. Previously the file watcher
would use a single thread. Many thread pool threads end up processing
these events concurrently, using resources, and the build would never
seem to end. Previously with the single thread processing the file
creation events the build would finish and the events would continue
to be process afterwards.

To avoid this problem the IDE now checks to see if the file is
excluded when looking to see if the file should be added to the
project. This avoids the slower check that is done when seeing if the
file should be included for a particular file glob. Before the exclude
check was done after it was determined the file was included in the
file glob. For ASP.NET Core projects the files in the node_modules
directory are excluded by imported file globs.

Fixes VSTS #835962 - Cannot stop SmartHotel360 solution build