[Ide] Support project template visibility with new templating engine
The old template engine supported a way for a template to indicate
if it should be visible if a project was being added to an existing
solution. This was not available when using the new templating engine
but can now be specified using the 'visibility' attribute for the
template in the .addin.xml file.

Fixes VSTS #839602 - Support project template visibility with new
templating engine