[VersionControl] Fix hang when opening a file
Selecting an item in the Solution pad would query an ignore-cache
version info every time.

Once for every command, we would create a task, spin up querying to native
and then output the result blocking wait.

Using the reading/writing task schedulers we have, we could end up in a potential
hang. And obviously, this was inefficiently recomputing everything.

Fixes VSTS #841060 - Hang when opening a file and lots of background threads