[Ide] Fix recent project timestamp being shown as in the future
If a project is marked as a favourite but a match cannot be found in
the recently used projects then the project gets a timestamp of
DateTime.Now. In the Get to Code dialog this resulted in the project
showing something similar to '59 minutes from now' and the tooltip
text showing a timestamp in the future.

The problem was that a UTC date is used for recently used projects
that are read back from the cached information but if a favourite
project is not found in this list a local date time is used. So code
that uses the DesktopService.RecentFiles could not rely on all the
timestamps being in UTC.

Fixes VSTS #843470 - Get to Code dialog shows project timestamps in
the future