[NuGet] Avoid loading type system info when restoring
When a NuGet restore is being run a task is set for the restore in
progress. The type system service now checks this task to see if it
can make any calls into the project model that use MSBuild. This limits
the overlap between the type system getting MSBuild information and
the NuGet packages being restored. Note that this is not perfect -
since the type system load can start before the NuGet restore has
started - so a check is made after each msbuild call to see if the
NuGet restore is started which will then pause the project load.

Fixes VSTS #844560 - Avoid loading TypeSystem when restoring NuGet
packages