[NuGet] Fix DotNetCliToolReferences not being restored
DotNetCliToolReferences are available in the dependency graph but are
treated as separate projects in this graph. Since these did not map
to an existing project they were not added to the full dependency
graph which resulted in these tools not being restored. Note that
these are only restored when the entire solution is restored.
Restoring a single project only restores the project itself not
the dotnet cli tool project in the dependency graph. MSBuild
supports restoring DotNetCliToolReferences in any project type that
uses PackageReferences so we also support this.

Fixes VSTS #567594 - DotNetCliToolReference is not restored