[Core] Fix old renamed file available in type system information.
Renaming a file in the project would result in the old filename and
the new filename being used as source files in the type system's
project information. The problem was that the cached source files
from running the CoreCompileDependsOn targets is not updated after
the rename so the old information is returned, which includes the
old filename. This cache is now cleared at the same time the evaluated
items is cleared.

Fixes VSTS #849821 - Project (hacky) cache stores duplicate build
actions