[Core] Use empty string for new file with missing item def properties
Adding a new file to the project that had a corresponding item
definition would add the file to the project in memory without the
metadata properties from the item definition, as expected, but the
project file would not reflect this. On reloading the project in
the IDE this new project file would then get those item definition
properties. Now if the new project file added is missing metadata
compared with the item definition then an empty string will be used
so the in memory information matches the msbuild item written to
the project file.