[Core] Improve item definition support when saving sdk style projects
Adding a new file that has an associated item definition to an sdk
style project was not removing metadata properties if they existed in
the item definition on saving. Also empty metadata properties were not
being added to the project file on saving if the new project item
added to the project did not have the corresponding item definition
properties.