Fix legacy editor assertion
The text editor line collection resets itself when there is a change in the
buffer, and there is an assertion that verifies that the snapshot assigned
to the collection is the same as the current snapshot of the editor.

There is however a case in which the editor snapshot is changed but
the change event is not raised. That may happen when the content type
changes. We were not handling that case, so the editor was asserting
since we were not accounting for the snapshot change.

This commit fixes the issue by subscribing the ContentTypeChanged event
and updating the cache when that happens.