Fixes https://devdiv.visualstudio.com/DevDiv/_workitems/edit/854282 - Open file isn't added to the workspace after project reload
This is a workaround fix for the symptom where when an open file is added to the project, it is not added to the workspace.

I notice that we issue UpdateRegisteredOpenDocuments () after we call Workspace.OnProjectReloaded, but in this code path it wasn't called.

Adding this call seems to add a case where TryOpenDocumentInWorkspace was called where project.ParentSolution was null. It seems innocuous to just check for that and exit early.