[NuGet] Add PackageReference metadata for development dependencies
Installing a NuGet package that is a development dependency, such as
GitInfo, would not add the PrivateAssets nor the IncludeAssets
metadata to the PackageReference. Now this is supported and mirrors the
behaviour of 'dotnet add package' and Visual Studio on Windows.

Fixes VSTS #634988 - PackageReference metadata not added for development
dependencies