[Ide] Close the New Project Dialog right after creation
The new wproject dialog can be closed once all template
bits have been processed and there is no way back.

This avoids collisions with other dialogs and alerts which
might pop up when the new project is loaded for the
first time, expecting the main IDE window to be visible
and not blocked.

Fixes VSTS #854324