Fix display issue with document tabs
When switching from Split to Tabs view in some cases the content from two
tabs are shown overlapped. This seems to be a cocoa/gtk integration issue.
For some reason the Notebook that hosts all the view is not properly hiding
pages that are not active.

As a workaround, the Notebook widget has been replaced by a simple
EventBox in which the active view is shown. This solves the problem and
simplifies the code, since we don't really use all the functionality of
Notebook (we use our custom tabstrip).

Fixes VSTS #848561 - Source view can't display correctly once switch from
"Split" View to “Source”