[NuGet] Update to NuGet 5.0.2
Fixes problem with package sources indicating that packages are signed
and Mono not supporting checking the package is signed resulting in
a failure to install or restore NuGet packages in projects that use
PackageReferences where the package is not in the local cache.
Fixes errors similar to:

Package 'Microsoft.AspNetCore.Authentication.Abstractions 2.1.0' from
source 'https://api.nuget.org/v3/index.json': This repository indicated
that all its packages are repository signed; however, this package is
unsigned.

Fixes VSTS #855951 - Package restores and builds failing with
"This repository indicated that all its packages are repository signed;
however, this package is unsigned."