[Ide] Move OpenDocument synchronizing logic to run in MonoDevelopWorkspace rather than NavigationService
The DocumentNavigationService is not the only user of the OpenDocument API, there are also code fixes
which sequence operations.

One example is suppressing code issues, where the chain of operations is:
* Apply fix
* Open document
* Navigate document to line

Without synchronously waiting for the document to open in the main OpenDocument code, we would trigger
the same behaviour as with jumping to declaration, and that would mean that the document might not be open
by the time it's requested to navigate to it.