[NuGet] Support partial installs into multi-target projects
A NuGet package being installed into a multi-target framework project
may not support all frameworks. When this happens the PackageReference
will now be added with a condition so the NuGet package is not used
for frameworks that are not supported.

<Project Sdk="Microsoft.NET.Sdk">

  <PropertyGroup>
    <TargetFrameworks>net472;netstandard1.2</TargetFrameworks>
  </PropertyGroup>

  <ItemGroup Condition="'$(TargetFramework)' == 'net472'">
    <PackageReference Include="EntityFramework" Version="6.2.0" />
  </ItemGroup>
</Project>

Note that NuGet does not support updating a NuGet package with a
condition. Visual Studio on Windows also shows an error about the NuGet
package being incompatible on updating it if it only supports certain
frameworks.

Fixes VSTS #572317 - Make adding PackageReferences to multi-framework
projects robust