Fix method insertion issue
When inserting a method in CodeGenerationService, wait for the ITextView to
be available before trying to do anything else with the content. Also, move the
caret to the new method once inserted.

Added new method to Document for getting a content asynchronously. The method
will wait until the content is available.

Fixes VSTS #858105 - Add button's event failed