Fix several toolbox issues:
SourceEditorView.cs, CodeTemplateToolboxProvider.cs: In the old editor, use a
private toolbox provider for snippets. The global provider tries to get the
editor from the consumer, and that's not possible anymore.

ToolboxService.cs: Support views that provide several
IToolboxDynamicProvider instances.

Document.cs: Added GetContents(forActiveView) method. Fixed some issues
with existing GetContent() methods.

DocumentController.cs: The Document property is not set for a controller it
will now check for the parent controller.

DocumentView.cs, DocumentViewContainer.cs: When a view gets the focus, make
sure it is set as active. This is important when in Split mode, since now
clicking in the different views of the split will set it as active.

Also fixed some focus issues with tab views. When switching between tabs,
ensure the activated tab gets the focus. Also fixed issue in Shown event, that
was not properly invoked for view containers. Added unit tests.